Move `QuestionSummarizeWidget` component
