+++ frontend/src/metabase/query_builder/components/view/ViewHeader/ViewHeader.jsx
@@ -18,7 +18,6 @@ import { navigateBackToDashboard } from "metabase/query_builder/actions";
 import { MODAL_TYPES } from "metabase/query_builder/constants";
 import { getDashboard } from "metabase/query_builder/selectors";
 import { FilterHeaderToggle, FilterHeader } from "../QuestionFilters";
-import { QuestionSummarizeWidget } from "../QuestionSummaries";
 import { ConvertQueryButton } from "./components/ConvertQueryButton";
 import { HeadBreadcrumbs } from "./components/HeaderBreadcrumbs";
 import { FilterHeaderButton } from "./components/FilterHeaderButton";
@@ -27,6 +26,7 @@ import { QuestionActions } from "./components/QuestionActions";
 import { QuestionNotebookButton } from "./components/QuestionNotebookButton";
 import { QuestionDataSource } from "./components/QuestionDataSource";
 import { QuestionDescription } from "./components/QuestionDescription";
+import { QuestionSummarizeWidget } from "./components/QuestionSummarizeWidget";
 import {
   AdHocViewHeading,
   SaveButton,
