+++ frontend/src/metabase/query_builder/components/view/ViewHeader/components/QuestionSummarizeWidget.jsx
@@ -3,8 +3,8 @@ import { t } from "ttag";
 
 import * as Lib from "metabase-lib";
 import { color } from "metabase/lib/colors";
-import ViewButton from "./ViewButton";
-import { HeaderButton } from "./ViewHeader/ViewHeader.styled";
+import ViewButton from "../../ViewButton";
+import { HeaderButton } from "../ViewHeader.styled";
 
 export function QuestionSummarizeWidget({
   isShowingSummarySidebar,
